patch: move 'types' files to '@types' and update imports.
