+++ src/@types/Reader.d.ts
