+++ src/app/(webapp)/(projects)/novels/[slug]/page.tsx
@@ -8,7 +8,7 @@ import { NovelData } from '@/components/project/NovelData';
 import { ProjectData } from '@/components/project/ProjectData';
 import { usePublicNovelSlug } from '@/hooks/usePublicApi';
 // import { useNovelStore } from '@/store/useNovelStore';
-import { TStatusNovel } from '@/types/System';
+import { TStatusNovel } from '@/@types/System';
 
 export default function Novel() {
   const pathname = usePathname();
