+++ src/components/common/checkbox/Checkbox.tsx
@@ -1,5 +1,5 @@
 import { cn } from '@/helpers/twUtils';
-import { ICheckbox } from '@/types/Checkbox';
+import { ICheckbox } from '@/@types/Checkbox';
 
 export const Checkbox = ({
   label,
