+++ src/components/common/tab/NavTabs.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement, useState, useCallback } from 'react';
 
 import { cn } from '@/helpers/twUtils';
-import { ITab } from '@/types/Tab';
+import { ITab } from '@/@types/Tab';
 
 interface CustomChildrenProps extends ReactElement, ITab {
   /* Nada */
